Verificando Owner e Permissions

+++ app/Http/Controllers/ProjectController.php
@@ -67,7 +67,7 @@ class ProjectController extends Controller
      */
     public function show($id)
     {
-        if ($this->checkProjectOwner($id) == false) {
+        if ($this->checkProjectPermissions($id) == false) {
             return ['error' => 'Forbidden'];
         }
         return $this->repository->find($id);
@@ -104,6 +104,11 @@ class ProjectController extends Controller
      */
     public function destroy($id)
     {
+
+        if ($this->checkProjectPermissions($id) == false) {
+            return ['error' => 'Access Forbidden'];
+        }
+
         if ($this->repository->delete($id)) {
             return ['success' => true];
         }
@@ -115,5 +120,21 @@ class ProjectController extends Controller
 
         return $this->repository->isOwner($projectId, $userId);
 
+    }
+
+    private function checkProjectMember($projectId) {
+
+        $userId =  \Authorizer::getResourceOwnerId();
+
+        return $this->repository->hasMember($projectId, $userId);
+
+    }
+
+    private function checkProjectPermissions ($projectId) {
+         if ($this->checkProjectOwner($projectId) or $this->checkProjectMember($projectId)) {
+            return true;
+         }
+
+        return false;
     }
 }
